# Replace `image` crate jpeg decoder with zune-jpeg