--- a/crates/re_components/Cargo.toml
+++ b/crates/re_components/Cargo.toml
@@ -30,7 +30,7 @@ ecolor = ["dep:ecolor"]
 glam = ["dep:glam"]
 
 ## Integration with the [`image`](https://crates.io/crates/image/) crate.
-image = ["dep:ecolor", "dep:image"]
+image = ["dep:ecolor", "dep:image", "dep:zune-core", "dep:zune-jpeg"]
 
 ## Enable (de)serialization using serde.
 serde = ["dep:serde", "half/serde", "re_log_types/serde"]
@@ -62,11 +62,11 @@ uuid = { version = "1.1", features = ["serde", "v4", "js"] }
 # Optional dependencies:
 ecolor = { workspace = true, optional = true }
 glam = { workspace = true, optional = true }
-image = { workspace = true, optional = true, default-features = false, features = [
-  "jpeg",
-] }
+image = { workspace = true, optional = true, default-features = false }
 rand = { version = "0.8", optional = true }
 serde = { version = "1", optional = true, features = ["derive", "rc"] }
+zune-core = { version = "0.2", optional = true }
+zune-jpeg = { version = "0.3", optional = true }
 
 # Native dependencies:
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]