--- a/rerun_py/Cargo.toml
+++ b/rerun_py/Cargo.toml
@@ -56,10 +56,7 @@ re_viewport.workspace = true
 
 arrow2 = { workspace = true, features = ["io_ipc", "io_print"] }
 document-features = "0.2"
-image = { workspace = true, default-features = false, features = [
-  "jpeg",
-  "png",
-] }
+image = { workspace = true, default-features = false, features = ["png"] }
 itertools = { workspace = true }
 mimalloc = { workspace = true, features = ["local_dynamic_tls"] }
 numpy = { version = "0.19.0", features = ["half"] }