@teh-cmc 2023-06-12
Seeing similar behavior on my _very_ different machine: ~14ms on `main` vs. ~8ms on this branch.