@emilk 2023-06-12
Worth noting: this slows down debug builds :(

`zune-jpeg` is 3x slower in debug builds than in release builds, even though we optimize all our dependencies even in debug builds with this:

https://github.com/rerun-io/rerun/blob/4acfcbd0d62da66f331d646a66bce50d038def02/Cargo.toml#L122-L124

This is strange